Neighbors code is ready for testing.
Includes code to finding all contact atoms, then for reducing the contact
atoms down to contact residues (or even nearest member of one group of atoms
to another group of atoms.) Definitely needs to be tested thoroughly, both
to make sure it runs on arbitrary boxes and to test for correctness. Plus,
I'm not sure this will work with multiple frames yet (and I think I want to
refactor a bit to make multiple frames work nicely).